+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 _Changes in the next release_
 
+### Changed
+- Externalize language strings for translations with Crowdin ([#27](https://github.com/unfoldedcircle/integration-denonavr/issues/27)))
+
 ---
 
 ## v0.8.3 - 2025-06-04